feat(webapp): compare against an uploaded report, and German guides
Work in progress from a parallel session, committed together on request.

- webapp/comparisons.py and webapp/imports.py accept a previously exported
  JSON or CSV report as the earlier side of /compare, with its own store,
  token and purge path; ADR 0057 records why an uploaded report is evidence
  rather than a scan.
- docs/de/ and frontend/templates/docs/de/ add reviewed German sources for
  the public guides; ADR 0058 records the rule they follow.
- The four string catalogues are rewritten to shorter, plainer wording, and
  the guides above are re-edited to match.

Also, so that the tree lints and its generated files are current: re.S becomes
re.DOTALL in tests/test_frontend_documentation.py, round() replaces a
redundant int(round()) in webapp/badge.py, webapp/feeds.py dates a schedule
with an aware UTC date rather than date.today(), and the generated
documentation is rebuilt.

Six tests fail on this commit, all of them prose assertions against the
rewritten catalogues - including the promise that every page says the check is
"not exhaustive". They are listed in the pull request discussion rather than
fixed here, because which side is right is the rewrite's decision to make.

+++ webapp/badge.py
@@ -46,7 +46,7 @@ _PADDING = 9
 
 
 def _segment_width(text: str) -> int:
-    return int(round(len(text) * _CHARACTER_WIDTH)) + 2 * _PADDING
+    return round(len(text) * _CHARACTER_WIDTH) + 2 * _PADDING
 
 
 def colour_for(rating: object) -> str:
